+++ lib/pippi/check_set_mapper.rb
@@ -28,6 +28,9 @@ module Pippi
           "ReverseFollowedByEach",
           "SelectFollowedBySelect"
         ],
+        "rails" => [
+          "StripFollowedByEmpty"
+        ],
         "training" => [
         ],
         "research" => [