+++ lib/pippi/checks/reverse_followed_by_each.rb
@@ -16,7 +16,7 @@ module Pippi::Checks
           # Ignore Array subclasses since reverse or each may have difference meanings
         else
           result.singleton_class.class_eval { prepend MyEach }
-          self.class._pippi_check_reverse_followed_by_each.array_mutator_methods.each do |this_means_its_ok_sym|
+          self.class._pippi_check_reverse_followed_by_each.mutator_methods.each do |this_means_its_ok_sym|
             result.define_singleton_method(this_means_its_ok_sym, self.class._pippi_check_reverse_followed_by_each.its_ok_watcher_proc(MyEach, :each))
           end
         end
@@ -37,7 +37,7 @@ module Pippi::Checks
 
     class Documentation
       def description
-        "Don't use each followed by reverse; use reverse_each instead"
+        "Don't use reverse followed by each; use reverse_each instead"
       end
 
       def sample